Don't suppress errors when reporting operator status
Previously errors applying updates to the operator's ClusterOperator
status were silently ignored.  This is mostly a problem when there's
an error reporting the Available status, because there is no immediate
retry in that case, which leaves the CVO unaware of the operator's
success.  This changes the reporter to always log errors applying the
updates, and to retry in the case the case of an error applying the
Available status.