Bug 1861642: Add maxNodeProvisionTime for baremetal
In baremetal environments it can easily take longer than the CA
default of 15mins for the node to become active after a scale-out
action.

The CA supports --max-node-provision-time[1] so adding support to
enable configuration of that value should allow tuning of the
time such that it's more suited to baremetal.

Accepted review suggestion to fix bug in regex
Co-authored-by: Joel Speed <Joel.speed@hotmail.co.uk>

[1] https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/FAQ.md#what-are-the-parameters-to-ca