install/90_alertrules: Drop ClusterAutoscalerOperatorDown
The cluster-version operator is responsible for complaining if the
autoscaler operator's deployment is sad, so no need for the operator
to handle this directly (we end up doubling up if there's an issue).
This drops the alert which was originally added in e1d802526a (Add
autoscaler down prometheus rule, 2019-11-05, #119).