add logic to detect GPU capacity and update accordingly
This change adds several functions to test whether a MachineSet has
capacity for GPUs on the Machines, and then applies a GPU accelerator
label, where appropriate, to the MachineSet. This label will help the
autoscaler to utilize its GPU custom node processor which will force the
autoscaler to wait for GPU drivers to be deployed before adding more
nodes. This alleviates a condition whereby the autoscaler can create too
many nodes for GPU workloads.