cleanup Makefile and add BUILD_IMAGE variable
This change removed the `GO111MODULE` variable from the Makefile and
adds a `BUILD_IMAGE` variable as a convenience when using alternate
build environments. The go module environment variable is being remove
because we the default behavior is now common for go tools and the old
behavior can be gained using the `-mod` flag to the go commands.

This change also cleans up the `DOCKER_CMD` for non-container builds to
include the same environment variables as the containerized targets.