revert GPU label changes from pr 223
this change is reverting the addition of the GPU label to MachineSet
objects as part of the CAO's GPU identification. It is being reverted
for a few reasons: the label is added to the wrong place on the
MachineSet, the label does not have the correct value, the label is not
being processed by the cluster autoscaler, and we have decided to not
have this bug fix automatically change values for the user.

This change does not have an impact on user experience as this value was
never being processed by the cluster autoscaler and should not pose a
risk to any user who is migrating older MachineSets as the label is
effectively a no-op.

ref: https://github.com/openshift/cluster-autoscaler-operator/pull/223
ref: https://bugzilla.redhat.com/show_bug.cgi?id=1943194