update GPU resource limits type to have validation
this change is part of the process to fix the GPU min and max resource
limits bug. The values for the GPUS.Type field should be used as a label
value on the nodes where gpus will be available. Because of this, users
will need to add those labels to their nodes, and as such the values for
the Type field needs to validated as if it were a label in the metadata
field. If the field is invalid a warning is created instead of producing
an error so that we do not break users with existing ClusterAutoscaler
resources.

ref: https://bugzilla.redhat.com/show_bug.cgi?id=1943194