add ValidatorResponse to util package
this change adds a new struct type to the util package that can be used
with the validators. It allows the return of both warnings and errors in
a single return value with a function to determine validity. The current
validators are updated and tests are added for the new struct.