Provide the architecture of the control plane as argument to --scale-up-from-zero-default-arch
When autoscale from zero is enabled in non-amd64 single-arch clusters on cloud providers managed by actuators that do not set the architecture in the capacity annotations (ref. MIXEDARCH-129), passing the argument --scale-up-from-zero-default-arch will allow not to default on the amd64 architecture for building the generic labels to scale out nodegroups from zero.

This commit sets the argument for the cluster-autoscaler to runtime.GOARCH.