vendor provisioning request api code
this change adds the cluster autoscaler api package to the tools package
so that the transitive dependency for the api code can be brought in to
the vendor directory. this will be used in following commits to generate
the manifest for the custom resource definition.

for this version of the vendor update, we are using a manually
calculated pseudo-version number so that we can get the api commit
associated with the 1.32 kubernetes release branch.