OCPBUGS-57041: update cluster provider type early
This bug was happening because the platform type of the cluster was not updated
before CreateAutoscaler is being called. This results in some arguments not getting
added to the deployment spec initially, but during a subsequent quick reconcile,
the platform type is updated in the UpdateAutoscaler function. So immediately,
the args are now added to the deployment, which triggers a new pod(s) rollout.

Signed-off-by: Max Cao <macao@redhat.com>